+++ projects/komponentkartan/src/lib/controls/titleValue/titleValue.angular.spec.ts
@@ -1,9 +1,7 @@
-
 import { ComponentFixture, TestBed, async, fakeAsync } from '@angular/core/testing';
 import { DebugElement, Component } from '@angular/core';
 import { BrowserDynamicTestingModule, platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';
 import { CommonModule } from '@angular/common';
-import { By } from '@angular/platform-browser';
 import { BrowserAnimationsModule, NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { IconModule } from '../icon/icon.module';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
@@ -12,11 +10,11 @@ import { TitleValueLayoutComponent } from './titleValueLayout.component';
 import { TitleValueComponent } from './titleValue.component';
 import { TitleValueContainerComponent } from './titleValueContainer/titleValueContainer.component';
 import { TitleValueHeadingComponent } from './titleValueHeading/titleValueHeading.component';
-import { InputComponent } from '../input/input.component';
 import { FormsModule } from '@angular/forms';
 import { IconComponent } from '../icon/icon.component';
 import { ErrorMessagePipe } from '../../pipes/errorMessagePipe';
 import { TruncatePipe } from '../../pipes/truncatePipe';
+import { InputComponent } from '../input/input.component';
 
 
 @Component({
@@ -26,19 +24,19 @@ import { TruncatePipe } from '../../pipes/truncatePipe';
         <vgr-title-value>
           <vgr-title-value-heading>Bruttobelopp</vgr-title-value-heading>
           <vgr-title-value-container>
-            <vgr-input [value]="10000" [suffix]="'kr'" [alignRight]="true"></vgr-input>
+            <vgr-input [value]="10000" [suffix]="'kr'" textAlign="right"></vgr-input>
           </vgr-title-value-container>
         </vgr-title-value>
         <vgr-title-value>
           <vgr-title-value-heading>Skattesats</vgr-title-value-heading>
           <vgr-title-value-container>
-            <vgr-input [value]="32" [suffix]="'%'" [alignRight]="true"></vgr-input>
+            <vgr-input [value]="32" [suffix]="'%'" textAlign="right"></vgr-input>
           </vgr-title-value-container>
         </vgr-title-value>
         <vgr-title-value>
           <vgr-title-value-heading>Nettobelopp</vgr-title-value-heading>
           <vgr-title-value-container>
-            <vgr-input [value]="32" [suffix]="'kr'" [alignRight]="true"></vgr-input>
+            <vgr-input [value]="32" [suffix]="'kr'" textAlign="right"></vgr-input>
           </vgr-title-value-container>
         </vgr-title-value>
       </vgr-title-value-layout>