+++ projects/komponentkartan/src/lib/komponentkartan.module.ts
@@ -51,7 +51,6 @@ const DEFAULT_PERFECT_SCROLLBAR_CONFIG: PerfectScrollbarConfigInterface = {
     type.ModalContentComponent,
     type.ModalHeaderComponent,
     type.ModalFooterComponent,
-    type.InputComponent,
     type.CardComponent,
     type.CardHeaderComponent,
     type.CardColumnComponent,
@@ -109,7 +108,8 @@ const DEFAULT_PERFECT_SCROLLBAR_CONFIG: PerfectScrollbarConfigInterface = {
     type.GridRowComponent,
     type.GridContentComponent,
     type.GridHeaderToolbarComponent,
-    type.NotificationComponent
+    type.NotificationComponent,
+    type.InputComponent
   ],
   exports: [
     type.SafePipe,
@@ -131,7 +131,6 @@ const DEFAULT_PERFECT_SCROLLBAR_CONFIG: PerfectScrollbarConfigInterface = {
     type.ModalContentComponent,
     type.ModalHeaderComponent,
     type.ModalFooterComponent,
-    type.InputComponent,
     type.CardComponent,
     type.CardHeaderComponent,
     type.CardColumnComponent,
@@ -189,7 +188,8 @@ const DEFAULT_PERFECT_SCROLLBAR_CONFIG: PerfectScrollbarConfigInterface = {
     type.GridRowComponent,
     type.GridContentComponent,
     type.GridHeaderToolbarComponent,
-    type.NotificationComponent
+    type.NotificationComponent,
+    type.InputComponent
   ],
   providers: [
     type.ModalService,