+++ src/app/search-results/search-results.component.ts
@@ -17,14 +17,14 @@ export class SearchResultsComponent implements OnInit {
   dropdownVisible = false;
   dropdownVisible_e1 = false;
   dropdownVisible_e2 = false;
-  filterBoxValue_e1: string;
+  filterBoxValue_e1 = '';
   filterBoxValue_e2: string;
   searchDescription_e1 = null;
   htmlExample1;
 
   constructor(htmlEncoder: HtmlEncodeService) {
     this.htmlExample1 = htmlEncoder.prepareHighlightedSection(`<div class="search-result-wrapper">
-    <vgr-filter-textbox (keydown)="filterSearch($event)" [value]="filterBoxValue_e1"></vgr-filter-textbox>
+    <vgr-input type="search" (keydown)="filterSearch($event)" [value]="filterBoxValue_e1"></vgr-input>
     <vgr-search-result [items]="filteredItems" [visible]="dropdownVisible_e1" maxItems="15" [description]="searchDescription_e1" (itemClick)="setResult($event)"></vgr-search-result>
   </div>`, 'html');
   }